better .ssh/ setup handling

+++ docker-entrypoint.sh
@@ -16,10 +16,15 @@ then
 fi
 
 # to be ok with git clone with ssh auth
-cp -f /root/.ssh/id_rsa.orig /root/.ssh/id_rsa
-cp -f /root/.ssh/id_rsa.pub.orig /root/.ssh/id_rsa.pub
-chmod 600 /root/.ssh/id_rsa
-chmod 644 /root/.ssh/id_rsa.pub
+if [ -f /root/.ssh/id_rsa.orig ]; then
+  cp -f /root/.ssh/id_rsa.orig /root/.ssh/id_rsa
+  cp -f /root/.ssh/id_rsa.pub.orig /root/.ssh/id_rsa.pub
+  chmod 600 /root/.ssh/id_rsa
+  chmod 644 /root/.ssh/id_rsa.pub
+  echo "-> SSH key found for git clone (.ssh/id_rsa is ready)"
+else
+  echo "-> No SSH key found for git clone (skip .ssh/ setup)"
+fi
 
 # run create-htmlreport.periodically.sh in the background
 # and log stdout/stderr to a file AND to stdout (in order to see the logs from docker)
